requirements.txt
pyyaml>=6.0
requests>=2.33
pySigma>=1.3
pySigma-backend-elasticsearch>=2.0
pySigma-pipeline-windows>=2.0
streamlit>=1.35
pandas>=2.0
plotly
pre-commit>=3.7
